- Changed: Some experimental changes to reduce the memory footprint of Game objects.  For this first modification changed regen ticks, race, and speed from int to short.  These few changes reduce the Storm footprint by 2K.  All of these may be changed to ushort but at this time I'm not sure if any of the values can be negative.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@2410 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/commands/gmcommands/npc.cs
@@ -157,10 +157,10 @@ namespace DOL.GS.Commands
 							return;
 						}
 
-						int speed = 200;
+						short speed = 200;
 						if (args.Length == 4)
 						{
-							speed = Convert.ToInt32(args[3]);
+							speed = Convert.ToInt16(args[3]);
 						}
 
 						int X = 0;
